Move expression builder to expression folder
