+++ src/parser/expression-parser.ts
@@ -7,10 +7,12 @@ import {
 import { AliasNode } from '../operation-node/alias-node.js'
 import { isOperationNodeSource } from '../operation-node/operation-node-source.js'
 import { OperationNode } from '../operation-node/operation-node.js'
-import { ExpressionBuilder } from '../query-builder/expression-builder.js'
+import {
+  expressionBuilder,
+  ExpressionBuilder,
+} from '../expression/expression-builder.js'
 import { SelectQueryBuilder } from '../query-builder/select-query-builder.js'
 import { isFunction } from '../util/object-utils.js'
-import { createExpressionBuilder } from './parse-utils.js'
 
 export type ExpressionOrFactory<DB, TB extends keyof DB, V> =
   // SQL treats a subquery with a single selection as a scalar. That's
@@ -44,7 +46,7 @@ export function parseExpression(
   if (isOperationNodeSource(exp)) {
     return exp.toOperationNode()
   } else if (isFunction(exp)) {
-    return exp(createExpressionBuilder()).toOperationNode()
+    return exp(expressionBuilder()).toOperationNode()
   }
 
   throw new Error(`invalid expression: ${JSON.stringify(exp)}`)
@@ -56,7 +58,7 @@ export function parseAliasedExpression(
   if (isOperationNodeSource(exp)) {
     return exp.toOperationNode()
   } else if (isFunction(exp)) {
-    return exp(createExpressionBuilder()).toOperationNode()
+    return exp(expressionBuilder()).toOperationNode()
   }
 
   throw new Error(`invalid aliased expression: ${JSON.stringify(exp)}`)
