+++ src/parser/parse-utils.ts
@@ -4,7 +4,7 @@ import { SelectQueryNode } from '../operation-node/select-query-node.js'
 import {
   ExpressionBuilder,
   ExpressionBuilderImpl,
-} from '../query-builder/expression-builder.js'
+} from '../expression/expression-builder.js'
 import { JoinBuilder } from '../query-builder/join-builder.js'
 import { OverBuilder } from '../query-builder/over-builder.js'
 import { SelectQueryBuilder } from '../query-builder/select-query-builder.js'
@@ -22,12 +22,6 @@ export function createSelectQueryBuilder(): SelectQueryBuilder<any, any, any> {
     queryId: createQueryId(),
     executor: NOOP_QUERY_EXECUTOR,
     queryNode: SelectQueryNode.create(parseTableExpressionOrList([])),
-  })
-}
-
-export function createExpressionBuilder(): ExpressionBuilder<any, any> {
-  return new ExpressionBuilderImpl({
-    executor: NOOP_QUERY_EXECUTOR,
   })
 }
