+++ src/parser/update-set-parser.ts
@@ -1,9 +1,8 @@
 import { ColumnNode } from '../operation-node/column-node.js'
 import { ColumnUpdateNode } from '../operation-node/column-update-node.js'
-import { ExpressionBuilder } from '../query-builder/expression-builder.js'
+import { expressionBuilder, ExpressionBuilder } from '../expression/expression-builder.js'
 import { UpdateKeys, UpdateType } from '../util/column-type.js'
 import { isFunction } from '../util/object-utils.js'
-import { createExpressionBuilder } from './parse-utils.js'
 import { parseValueExpression, ValueExpression } from './value-parser.js'
 
 export type UpdateObject<DB, TB extends keyof DB, UT extends keyof DB = TB> = {
@@ -28,7 +27,7 @@ export function parseUpdateExpression(
   update: UpdateExpression<any, any, any>
 ): ReadonlyArray<ColumnUpdateNode> {
   const updateObj = isFunction(update)
-    ? update(createExpressionBuilder())
+    ? update(expressionBuilder())
     : update
 
   return Object.entries(updateObj)
